Updated test to verify that custom deployments work, even for deployment scripts that sit outside the repository itself.
Also added etag on successful delete requests as that allows a client to keep track of the latest etag when deleting as well as creating/updating content.

Fixed bug where we did not deploy content in case of a merge. A test bug prevented us from discovering it. The test bug has also been fixed.

Lastly, removed an extraneous row in exponential backoff numbers so we are back to 8 instead of 9.