Added a new setting to WebJobs called "is_in_place", setting this to true will result in **not** copying the WebJobs binaries to a temporary directory before running it, instead running it in the same directory that it exists.
Pros for using it:
* It's faster - reduce the time to copy to the temporary directory, significant if there are many files.
* Ensures it'll run, not failing on the long path issue when running from the temporary directory.

Cons:
* Will lock files that are executed (.exe / .dll) and won't allow redeploying them.

This feature is enabled by default for node.js WebJobs since there is less of a risk for locking files (.js files are not usually locked).

Also moved the WebJob caching to happen after a triggered job invocation responds with 202 to reduce the amount of time it takes to invoke a triggered WebJob.

Fixes #1376