Fix serviceSitePath and sitesPath for Release configuration.
Use XML transformation to have the keys in Web.config point to the correct paths. When building with Release configuration they are one level higher. This is because Release is used when deploying Kudu to a server, so Kudu.Services.Web\ and apps\ are located in e.g. C:\inetpub\ while KuduWeb\ is in e.g. C:\inetpub\wwwroot\.