Linux - Support for container restart trigger (#2529)
Linux - Support for container restart trigger

Linux App Service now supports a generic file-based communication mechanism
for requesting a restart of the app container. This was previously implemented
in a limited fashion only for the "Docker CI" webhook. This generalizes the
implementation and also requests a restart after a git deployment by default,
in addition to receipt of a Docker webhook request.