Fix for bug #2546
Suggesting fix for https://github.com/projectkudu/kudu/issues/2546
Edge is throwing the following error as dataTransfer property is null.

SCRIPT5007: Unable to get property 'items' of undefined or null reference
filebrowser.js (802,9)

The OR statement results in an error and the function terminates, hence the file upload fails. Checking for null and assuming zip to allow the function to continue. 

Tested the code change in Edge.