Support WAR file based deployments
This commit introduces a new API (/api/wardeploy) for publishing WAR files to Java Web Apps to address the deployment issues with scaled out Java apps.

Background:
Publishing a WAR file to a Java web app is a common scenario. The WAR file is published to the wwwroot/webapps directory. Tomcat monitors the wwwroot/webapps directory for the presence of new *.war files (or changes to existing ones). The moment it detects either of that, it unzips the war file into a new directory that matches the name of the war file and loads the unzipped web application from that directory (WAR files are just zip files with a war extension).

In a scaled out web app, multiple Tomcat instances (one on each worker) attempt to unzip the WAR file causing multiple issues. Here are some common problems: When a change to the WAR file is detected, multiple Tomcat instances attempt to lock and unzip it, causing locking problems. Another issues is Tomcat attempting to reload a webapp which is in the middle of being updated by another Tomcat instance resulting in the Tomcat instance going into a bad state. There are few more race conditions like these which eventually result in Tomcat responding to incoming requests with 404.

/api/wardeploy:
This commit builds on top of the existing zipdeploy mechanism to perform unzipping of WAR files outside the wwwwroot directory and deploys the expanded WAR file to the wwwwroot/webapps/ROOT directory which is monitored by Tomcat. After the deployment is complete wardeploy touches the wwwroot/webapps/ROOT/WEB-INF/web.xml file which causes Tomcat to reload the web app (equivalent to touching web.config). The race conditions with scaled out Java web apps are avoided as WAR file is not directly published to the wwwroot/webapps directory.